examples -> cli
