+++ cellarium/ml/cli/probabilistic_pca.py
@@ -62,10 +62,11 @@ class _LightningCLIWithLinks(LightningCLI):
         parser.link_arguments("data.n_vars", "model.module.init_args.g_genes", apply_on="instantiate")
 
 
-def main():
+def main(args=None):
     _LightningCLIWithLinks(
         TrainingPlan,
         DistributedAnnDataCollectionDataModule,
+        args=args,
     )
